requirements.txt
Django==4.2.6
python-dotenv==1.0.0
django-filter==23.3
djangorestframework==3.14.0
django-cors-headers==4.3.0
openai==0.28.1
